+++ spec/features/transfers/tasks/users_can_complete_all_conditions_met_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.feature "Users can complete the all conditions met task" do
+RSpec.feature "Users can complete the authority to proceed" do
   let(:user) { create(:user, :regional_delivery_officer) }
   let(:project) { create(:transfer_project, assigned_to: user) }
 
@@ -10,14 +10,14 @@ RSpec.feature "Users can complete the all conditions met task" do
     visit project_tasks_path(project)
   end
 
-  describe "the confirm all conditions met task" do
-    scenario "they can confirm all conditions have been met" do
+  describe "the authority to proceed task" do
+    scenario "they can confirm the transfer has authority to proceed" do
       visit project_tasks_path(project)
-      click_on "Confirm all conditions have been met"
+      click_on "Confirm this transfer has authority to proceed"
       page.find_all(".govuk-checkboxes__input").each { |checkbox| checkbox.click }
       click_on I18n.t("task_list.continue_button.text")
 
-      expect(project.reload.all_conditions_met).to be true
+      expect(project.reload.authority_to_proceed).to be true
     end
   end
 end